fix(#8533): move the sockaddr layout out of the root package into org.eolang.sys
