+++ eo-runtime/src/main/java/org/eolang/posix/AcceptSyscall.java
@@ -9,8 +9,8 @@ import org.eolang.Data;
 import org.eolang.Int;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
-import org.eolang.Sockaddr;
 import org.eolang.Syscall;
+import org.eolang.sys.Sockaddr;
 
 /**
  * Accept syscall.
