+++ eo-runtime/src/main/java/org/eolang/posix/ConnectSyscall.java
@@ -8,8 +8,8 @@ import org.eolang.Data;
 import org.eolang.Int;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
-import org.eolang.Sockaddr;
 import org.eolang.Syscall;
+import org.eolang.sys.Sockaddr;
 
 /**
  * Connect syscall.
