+++ eo-runtime/src/main/java/org/eolang/win32/ConnectFuncCall.java
@@ -10,8 +10,8 @@ import org.eolang.Dataized;
 import org.eolang.Int;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
-import org.eolang.SockaddrIn;
 import org.eolang.Syscall;
+import org.eolang.sys.SockaddrIn;
 
 /**
  * The 'connect' WS2_32 function call.
