+++ eo-runtime/src/test/java/org/eolang/sys/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
+ * SPDX-License-Identifier: MIT
+ */
+
+/**
+ * Tests for what the syscall adapters share.
+ * @since 0.77.0
+ */
+package org.eolang.sys;
